remove reorg
